Random graph generators
Added `kronecker` graph generator according to `Graph500`.

Also added a script that generates them.

+++ src/main/java/triangle_counting/GraphManager.java
@@ -93,7 +93,7 @@ public class GraphManager {
             
             arrayGraph[node] = new int[nodeNeighbors.size()];            
             for (int destNodeInd = 0; destNodeInd < nodeNeighbors.size(); destNodeInd++) {
-                arrayGraph[node][destNodeInd] = nodeNeighbors.get(destNodeInd).intValue();
+                arrayGraph[node][destNodeInd] = nodeNeighbors.get(destNodeInd);
             }
         }
         return arrayGraph;
